+++ src/utils/mod.rs
@@ -1,5 +1,7 @@
+mod fail;
 mod global_tracer;
 mod reporter;
 
+pub use self::fail::FailSpan;
 pub use self::global_tracer::GlobalTracer;
 pub use self::reporter::ReporterThread;